fix error prompt in project form page blocked by top navbars
